src/main/webapp/app/entities/tai-san-dgc/delete/tai-san-dgc-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { provideHttpClient, HttpResponse } from '@angular/common/http';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { TaiSanDgcService } from '../service/tai-san-dgc.service';

import { TaiSanDgcDeleteDialogComponent } from './tai-san-dgc-delete-dialog.component';

describe('TaiSanDgc Management Delete Component', () => {
  let comp: TaiSanDgcDeleteDialogComponent;
  let fixture: ComponentFixture<TaiSanDgcDeleteDialogComponent>;
  let service: TaiSanDgcService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [TaiSanDgcDeleteDialogComponent],
      providers: [provideHttpClient(), NgbActiveModal],
    })
      .overrideTemplate(TaiSanDgcDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(TaiSanDgcDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(TaiSanDgcService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
